# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CoachLink Platform API
  description: |
    API для платформы автоматизации тренировочного процесса
    и взаимодействия в системе «тренер — спортсмен».

    ## Аутентификация
    Все запросы (кроме /auth/*) требуют JWT access-token
    в заголовке `Authorization: Bearer <token>`.

    ## Роли
    - `coach` — тренер
    - `athlete` — спортсмен

    ## Формат дат
    - Даты в формате ISO 8601: `2026-03-28`
    - Timestamps в формате ISO 8601 с таймзоной: `2026-03-28T12:00:00Z`

    ## Пагинация
    Списочные эндпоинты поддерживают query-параметры `page` и `page_size`.
    Ответ содержит объект `pagination` с метаданными.
  version: 1.0.0
  contact:
    name: CoachLink Team
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
...
paths:
...
  /api/v1/training/plans:
    post:
      tags: [Training]
      summary: Создать тренировочный план и назначить спортсменам
      description: |
        Тренер создаёт план и назначает его:
        - Конкретным спортсменам (по `athlete_ids`)
        - Группе (по `group_id` — назначается всем спортсменам в группе)

        Должен быть указан хотя бы один из параметров: `athlete_ids` или `group_id`.
        Можно указать оба — тогда задание получат все указанные спортсмены
        плюс все спортсмены из группы (без дублирования).
      operationId: createTrainingPlan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrainingPlanRequest'
      responses:
        '201':
          description: План создан, задания назначены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTrainingPlanResponse'
        '400':
          description: Ошибка валидации (например, не указаны получатели)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/training/assignments:
    get:
      tags: [Training]
      summary: Список назначений
      description: |
        - **Тренер** видит выданные задания (не архивированные).
        - **Спортсмен** видит свои назначения со статусом `assigned`.

        Поддерживает фильтрацию и сортировку.
      operationId: getAssignments
      security:
        - BearerAuth: []
      parameters:
        - name: athlete_full_name
          in: query
          required: false
          description: Фильтр по ФИО спортсмена (подстрока, только для тренера)
          schema:
            type: string
        - name: athlete_login
          in: query
          required: false
          description: Фильтр по логину спортсмена (подстрока, только для тренера)
          schema:
            type: string
        - name: date_from
          in: query
          required: false
          description: Фильтр по дате начала (>=)
          schema:
            type: string
            format: date
            example: "2026-03-01"
        - name: date_to
          in: query
          required: false
          description: Фильтр по дате окончания (<=)
          schema:
            type: string
            format: date
            example: "2026-03-31"
        - name: status
          in: query
          required: false
          description: Фильтр по статусу (только для тренера)
          schema:
            type: string
            enum: [assigned, completed]
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum: [date_asc, date_desc, name_asc, name_desc]
            default: date_desc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список назначений
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssignmentListItem'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/training/assignments/archived:
    get:
      tags: [Training]
      summary: Архив назначений (для тренера)
      operationId: getArchivedAssignments
      security:
        - BearerAuth: []
      parameters:
        - name: athlete_full_name
          in: query
          required: false
          schema:
            type: string
        - name: athlete_login
          in: query
          required: false
          schema:
            type: string
        - name: date_from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Архив назначений
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssignmentListItem'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/training/assignments/{assignmentId}:
    get:
      tags: [Training]
      summary: Детали назначения
      description: Возвращает полную информацию о назначении включая описание тренировки.
      operationId: getAssignment
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: Детали назначения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет доступа к этому назначению
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Training]
      summary: Удалить назначение
      description: Тренер удаляет назначение. Если к нему был отчёт, отчёт тоже удаляется.
      operationId: deleteAssignment
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '204':
          description: Назначение удалено
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваше назначение
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/training/assignments/{assignmentId}/archive:
    put:
      tags: [Training]
      summary: Архивировать назначение
      description: |
        Тренер перемещает выполненное назначение в архив.
        Назначение должно иметь статус `completed` (отчёт отправлен).
      operationId: archiveAssignment
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: Назначение архивировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentListItem'
        '400':
          description: Назначение ещё не выполнено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваше назначение
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------
  # REPORTS
  # ----------------------------------------------------------
  /api/v1/training/assignments/{assignmentId}/report:
    post:
      tags: [Reports]
      summary: Отправить отчёт по заданию
      description: |
        Спортсмен отправляет отчёт о выполненной тренировке.
        После отправки статус назначения меняется на `completed`.
        Повторная отправка невозможна.
      operationId: submitReport
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Отчёт отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingReport'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только спортсменам / это не ваше задание
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Отчёт уже был отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Reports]
      summary: Просмотреть отчёт по заданию
      description: |
        Доступно:
        - Тренеру, который выдал задание
        - Спортсмену, который отправил отчёт
      operationId: getReport
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: Отчёт по заданию
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет доступа к этому отчёту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Назначение или отчёт не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ----------------------------------------------------------
  # TEMPLATES
  # ----------------------------------------------------------
  /api/v1/training/templates:
    post:
      tags: [Templates]
      summary: Создать шаблон тренировки
      operationId: createTemplate
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Шаблон создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Templates]
      summary: Список шаблонов тренера
      operationId: getTemplates
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Поиск по заголовку шаблона
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список шаблонов
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrainingTemplate'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/training/templates/{templateId}:
    get:
      tags: [Templates]
      summary: Детали шаблона
      operationId: getTemplate
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: Детали шаблона
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Это не ваш шаблон
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Templates]
      summary: Обновить шаблон
      operationId: updateTemplate
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateRequest'
      responses:
        '200':
          description: Шаблон обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Это не ваш шаблон
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Templates]
      summary: Удалить шаблон
      operationId: deleteTemplate
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '204':
          description: Шаблон удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Это не ваш шаблон
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------
  # NOTIFICATIONS
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access-token, полученный при /auth/login или /auth/refresh

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
...
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TemplateId:
      name: templateId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateTrainingPlanRequest:
      type: object
      required: [title, description, scheduled_date]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Заголовок тренировки
          example: "Развивающий кросс"
        description:
          type: string
          minLength: 1
          description: Подробное описание тренировки
          example: "Кросс 8 км в аэробной зоне (пульс 140-155). Разминка 2 км трусцой, основная часть 5 км в темпе 4:30-4:45/км, заминка 1 км."
        scheduled_date:
          type: string
          format: date
          description: Дата тренировки (ГГГГ-ММ-ДД)
          example: "2026-04-05"
        athlete_ids:
          type: array
          items:
            type: string
            format: uuid
          description: |
            Список ID спортсменов для назначения.
            Должен быть указан хотя бы один из: athlete_ids или group_id.
          example: ["550e8400-e29b-41d4-a716-446655440001", "550e8400-e29b-41d4-a716-446655440002"]
        group_id:
          type: string
          format: uuid
          description: |
            ID группы. Задание будет назначено всем спортсменам в группе.
            Должен быть указан хотя бы один из: athlete_ids или group_id.
        template_id:
          type: string
          format: uuid
          description: |
            ID шаблона. Если указан, title и description берутся из шаблона
            (но могут быть переопределены в запросе).
        save_as_template:
          type: boolean
          default: false
          description: Сохранить план как шаблон для будущего использования
...